backend/requirements.txt
Flask~=3.0.3
Flask-Cors~=5.0.0
Flask-SQLAlchemy~=3.1.1
Flask-Migrate~=4.0.7
psycopg2~=2.9.7
python-dotenv~=1.0.0
chardet~=5.2.0

SQLAlchemy~=2.0.35
alembic~=1.13.3
typing_extensions~=4.12.2
PuLP~=2.9.0
numpy~=2.1.1